Changed guid_hash() to return vpr::Uint16 to address run-time failures on CentOS 3.6/RHEL3.
git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/PyJuggler@18308 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ src/vpr/_GUID.cpp
@@ -75,7 +75,7 @@ struct GuidPickle : pickle_suite
    }
 };
 
-vpr::Uint32 guid_hash(vpr::GUID* g)
+vpr::Uint16 guid_hash(vpr::GUID* g)
 {
    static vpr::GUID::hash hash;
    return hash(*g);
